postreplies done

+++ src/blog/blog.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateBlogDto } from './dto/create-blog.dto';
 import { UpdateBlogDto } from './dto/update-blog.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -10,17 +10,16 @@ export class BlogService {
   constructor(@InjectRepository(Blog) private blogrepository:Repository<Blog>){}
 
 
-  async create(createBlogDto: CreateBlogDto,user):Promise<Blog> {
-
-    const newblog = this.blogrepository.create({
+  async create(createBlogDto: CreateBlogDto, user): Promise<Blog> {
+    const newBlog = this.blogrepository.create({
       ...createBlogDto,
-      createdby:user
+      createdby: user,
     });
-    return this.blogrepository.save(newblog) 
+    return this.blogrepository.save(newBlog);
   }
 
   findAll() {
-    return this.blogrepository.find()
+    return this.blogrepository.find();
   }
 
   async findOne(id: number): Promise<Blog> {
@@ -31,14 +30,17 @@ export class BlogService {
     return blog;
   }
 
-  async update(id: number, updateBlogDto: UpdateBlogDto): Promise<Blog> {
+  async update(id: number, updateBlogDto: UpdateBlogDto, user): Promise<Blog> {
     const blog = await this.findOne(id);
-    Object.assign(blog, updateBlogDto);
+    Object.assign(blog, updateBlogDto, { updatedBy: user });
     return this.blogrepository.save(blog);
   }
 
-  async remove(id: number): Promise<void> {
+  async remove(id: number, user): Promise<void> {
     const blog = await this.findOne(id);
+    if (blog.createdby !== user) {
+      throw new ForbiddenException('You do not have permission to delete this blog');
+    }
     await this.blogrepository.remove(blog);
   }
 }
